Rework the way level value it used inside `opt::option`.
Previously, `get_level` returned `std::uintmax_t(-1)` if value is engaged and other values as level number (`0` for first layer, `1` for second layer, etc.).
Currently, `get_level` can return any value in range of `std::uintmax_t`, but `0` will indicate that the contained value is empty (accordingly, all the others that it's not).

This will improve code generation because the compiler doesn't need to check if the level is in the `max_level` range but can just compare it to zero.